added some packet parsing logic when receiving from a client

+++ Sources/App/packets/PacketGameplayID.swift
@@ -7,5 +7,5 @@
 
 import Foundation
 
-public protocol PacketGameplayID : RawRepresentable where RawValue == UInt {
+public protocol PacketGameplayID : RawRepresentable where RawValue == UInt8 {
 }
